Refactor CNN and BiLSTM to BaseNN
